+++ test/unit/mongo_db_admin/unit_test_run.sh
@@ -12,11 +12,8 @@ echo "Unit test:"
 /usr/bin/python test/unit/mongo_db_admin/get_log.py
 /usr/bin/python test/unit/mongo_db_admin/help_message.py
 /usr/bin/python test/unit/mongo_db_admin/main.py
-/usr/bin/python test/unit/mongo_db_admin/process_dbs_tbls.py
 /usr/bin/python test/unit/mongo_db_admin/process_mail.py
-/usr/bin/python test/unit/mongo_db_admin/process_request.py
 /usr/bin/python test/unit/mongo_db_admin/rotate.py
-/usr/bin/python test/unit/mongo_db_admin/run_dbcc.py
 /usr/bin/python test/unit/mongo_db_admin/run_program.py
 /usr/bin/python test/unit/mongo_db_admin/status.py
 /usr/bin/python test/unit/mongo_db_admin/create_data_config.py